Move the functions for reading/writing memory to `util.rs`
